\name{gen3sis-package}
\alias{gen3sis-package}
\alias{gen3sis}
\docType{package}
\title{
\packageTitle{gen3sis}
}
\description{
\packageDescription{gen3sis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gen3sis}
\packageIndices{gen3sis}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gen3sis}

Maintainer: \packageMaintainer{gen3sis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
